+++ src/main/java/com/piggahbrostudios/elderscrollsmod/capabilities/StorageHandler.java
@@ -10,7 +10,7 @@ public class StorageHandler implements Capability.IStorage<IStorage> {
     @Override
     public NBTBase writeNBT(Capability<IStorage> capability, IStorage instance, EnumFacing side) {
         final NBTTagCompound tag = new NBTTagCompound();
-        tag.setInteger("magika", instance.getMagika());
+        tag.setFloat("magika", instance.getMagika());
         tag.setInteger("stamina", instance.getStamina());
         return tag;
     }
@@ -18,7 +18,7 @@ public class StorageHandler implements Capability.IStorage<IStorage> {
     @Override
     public void readNBT(Capability<IStorage> capability, IStorage instance, EnumFacing side, NBTBase nbt) {
         final NBTTagCompound tag = (NBTTagCompound) nbt;
-        instance.setMagika(tag.getInteger("magika"));
+        instance.setMagika(tag.getFloat("magika"));
         instance.setStamina(tag.getInteger("stamina"));
     }
